+++ lizard_connector/__init__.py
@@ -3,3 +3,4 @@ from lizard_connector import jsdatetime
 from lizard_connector import connector
 from lizard_connector import queries
 from lizard_connector import parsers
+from lizard_connector import test