+++ lizard_connector/connector.py
@@ -11,8 +11,11 @@ Includes:
 - Endpoints (Lizard api endoints)
 - Connector (http handling)
 """
+from __future__ import absolute_import
+from __future__ import print_function
+from __future__ import unicode_literals
+from __future__ import generators
 
-import functools
 import json
 
 import time
@@ -23,7 +26,9 @@ try:
     from urllib.parse import urljoin
     import urllib.request as urllib_request
     from urllib.request import urlopen
+    PYTHON_VERSION = 3
 except ImportError:
+    PYTHON_VERSION = 2
     from urllib import urlencode
     from urlparse import urljoin
     import urllib2 as urllib_request
@@ -67,42 +72,41 @@ def save_to_json(result, file_base="api_result"):
 
 class Connector(object):
 
-    def __init__(self, max_results=1000, username=None, password=None):
+    def __init__(self, username=None, password=None):
         """
         Args:
-            max_results (int): maximum number of results allowed from one get.
             username (str): lizard-api user name to log in. Without one no
                             login is used.
             password (str): lizard-api password to log in. Without one no login
                             is used.
         """
-        self.max_results = max_results
         self.username = username
         self.password = password
 
-    def get(self, url):
+    def get(self, url, raise_error_on_next_url=False):
         """
         GET a json from the api.
 
         Args:
             url (str): Lizard-api valid url.
-
+            raise_error_on_next_url (bool): when provided an error is raised
+                when a response contains a next_url field and the field is not
+                null.
         Returns:
             A list of dictionaries of the 'results'-part of the api-response.
         """
         json_ = self.perform_request(url)
-        self.handle_count(json_)
-        json_ = json_.get('results', json_)
-        return json_
-
-    def handle_count(self, json_):
-        count = json_.get('count')
-        if (count or 0) > self.max_results:
+        if raise_error_on_next_url and bool(json_.get('next', False)):
             raise LizardApiTooManyResults(
-                'Too many results: {} found, while max {} are accepted'.format(
-                    count or 0, self.max_results)
+                "\nThe Lizard API returns more than one result page. Please \n"
+                "use `download_paginated` or `download_async` methods \n"
+                "instead for large api responses. Or increase the page_size \n"
+                "in the request parameters."
             )
-        return count
+        try:
+            json_ = json_.get('results', json_)
+        finally:
+            return json_
 
     def post(self, url, data):
         """
@@ -138,12 +142,11 @@ class Connector(object):
                 url,
                 headers=headers,
                 data=json.dumps(data).encode('utf-8'),
-                method="POST")
+            )
         else:
             request_obj = urllib_request.Request(url, headers=self.header)
         resp = urlopen(request_obj)
         content = resp.read().decode('UTF-8')
-        # print(content)
         return json.loads(content)
 
     @property
@@ -173,17 +176,9 @@ class PaginatedRequest(object):
     def __init__(self, endpoint, url):
         """
         Args:
-            max_results (int): maximum number of results allowed from one get.
-            username (str): lizard-api user name to log in. Without one no
-                            login is used.
-            password (str): lizard-api password to log in. Without one no login
-                            is used.
-            paginate (bool): when set to True, paginate results. The connector
-                             now can be used as an iterator also the next_page
-                             method can be used to obtain the next page. When
-                             set to False a query is built with a max size of
-                             max_results. All these results are returned at
-                             once.
+            endpoint (Endpoint): Endpoint object.
+            url (str): First url to start the paginated request. This should
+            be Lizard-api valid url.
         """
         self._endpoint = endpoint
         self.next_url = url
@@ -200,7 +195,7 @@ class PaginatedRequest(object):
             A list of dictionaries of the 'results'-part of the api-response.
         """
         json_ = self._endpoint.perform_request(self.next_url)
-        self._count = self._endpoint.handle_count(json_)
+        self._count = json_.get('count')
         self.next_url = json_.get('next')
         json_ = json_.get('results', json_)
         return json_
@@ -230,40 +225,32 @@ class PaginatedRequest(object):
 
 
 class Endpoint(Connector):
-    max_results = 1000
 
     def __init__(self, endpoint, base="https://demo.lizard.net", **kwargs):
         """
         Args:
             base (str): lizard-nxt url.
-            max_results (int): maximum number of results allowed from one get.
             username (str): lizard-api user name to log in. Without one no
                             login is used.
             password (str): lizard-api password to log in. Without one no login
                             is used.
-            paginate (bool): when set to True, paginate results. The connector
-                             now can be used as an iterator also the next_page
-                             method can be used to obtain the next page. When
-                             set to False a query is built with a max size of
-                             max_results. All these results are returned at
-                             once.
         """
         super(Endpoint, self).__init__(**kwargs)
         self.endpoint = endpoint
         base = base.strip(r'/')
-        if not base.startswith('https'):
+        if not base.startswith('https') and not 'localhost' in base:
             raise InvalidUrlError('base should start with https')
         base = urljoin(base, 'api/v2') + "/"
         self.base_url = urljoin(base, self.endpoint) + "/"
 
-    def _build_url(self, *querydicts, **queries):
+    def _build_url(self, page_size=1000, *querydicts, **queries):
         q = lizard_connector.queries.QueryDictionary(
-            page_size=self.max_results)
+            page_size=page_size)
         q.update(*querydicts, **queries)
         query = "?" + urlencode(q)
         return urljoin(self.base_url, query)
 
-    def download(self, *querydicts, **queries):
+    def download(self, page_size=1000, *querydicts, **queries):
         """
         Query the api at this endpoint and download its data.
 
@@ -271,12 +258,14 @@ class Endpoint(Connector):
         Stores the api-response as a dict in the results attribute.
 
         Args:
+            page_size (int): the page_size parameter when more results are
+                returned than the page size allows an error is returned.
             querydicts (iterable): all key valuepairs from dictionaries are
                                    used as queries.
             queries (dict): all keyword arguments are used as queries.
         """
-        url = self._build_url(*querydicts, **queries)
-        return self.get(url)
+        url = self._build_url(page_size=page_size, *querydicts, **queries)
+        return self.get(url, raise_error_on_next_url=True)
 
     def download_paginated(self, page_size=100, *querydicts, **queries):
         """
@@ -295,9 +284,6 @@ class Endpoint(Connector):
             querydicts (iterable): all key valuepairs from dictionaries are
                                    used as queries.
             page_size (int): number of results per iteration.
-            lock (Lock): a threading lock
-            thread_class (class): either Thread or a similar Thread class like
-                                  QThread
             queries (dict): all keyword arguments are used as queries.
         Returns:
             an iterable that returns the results of each page.
@@ -369,7 +355,9 @@ class Endpoint(Connector):
             queries (dict): all keyword arguments are used as queries.
         """
         queries.update({"async": "true"})
-        task_url = self.download(*querydicts, **queries).get('url')
+        page_size = queries.pop('page_size', 0)
+        task_url = self.download(
+            page_size=page_size, *querydicts, **queries).get('url')
         keep_polling = True
         while keep_polling:
             result, keep_polling = self._poll_task(task_url)
@@ -393,95 +381,3 @@ class Endpoint(Connector):
         else:
             post_url = self.base_url
         return self.post(post_url, data)
-
-
-class Datatype(object):
-    """
-    Experimental class that forms the building block for data types.
-
-    Each data type contains a list of queries for each relevant lizard-api
-    endpoint. These are the only viable queries for each endpoint.
-    """
-
-    @property
-    def queries(self):
-        """
-        This property needs to be implemented for the class to function.
-        """
-        raise NotImplementedError
-
-    def endpoints(self):
-        """Lists all available endpoints for this datatype."""
-        return list(self.queries.keys())
-
-    def endpoint_queries(self, endpoint):
-        """Lists all available queries for an endpoint of this datatype."""
-        return list(self.queries[endpoint].keys())
-
-    def download(self, endpoint, *querydicts, **queries):
-        """Downloads a json as a dict from an endpoint."""
-        endpoint_ = Endpoint(endpoint)
-        return endpoint_.download(*querydicts, **queries)
-
-
-class Timeseries(Datatype):
-    queries = {
-        "timeseries": {
-            "in_bbox": functools.partial(
-                lizard_connector.queries.in_bbox, endpoint="timeseries"),
-            "distance_to_point": lizard_connector.queries.distance_to_point,
-            "datetime_limits": lizard_connector.queries.datetime_limits,
-            "organisation": functools.partial(
-                lizard_connector.queries.organisation, endpoint="timeseries"),
-            "statistics": lizard_connector.queries.statistics
-        },
-        "locations": {
-            "in_bbox": functools.partial(
-                lizard_connector.queries.in_bbox, endpoint="timeseries"),
-            "distance_to_point": lizard_connector.queries.distance_to_point,
-            "organisation": functools.partial(
-                lizard_connector.queries.organisation, endpoint="locations")
-        }
-    }
-
-
-class Rasters(Datatype):
-    queries = {
-        "rasters": {
-            "feature_info": lizard_connector.queries.feature_info,
-            "limits": lizard_connector.queries.limits,
-            "organisation": functools.partial(
-                lizard_connector.queries.organisation, endpoint="rasters")
-        }
-    }
-
-
-class Events(Datatype):
-    queries = {
-        "events": {
-            "organisation": functools.partial(
-                lizard_connector.queries.organisation, endpoint="events")
-        }
-    }
-
-
-class Assets(Datatype):
-    queries = {
-        "leveerings": {
-            "organisation": functools.partial(
-                lizard_connector.queries.organisation, endpoint="leveerings")
-        },
-        "levees": {
-            "organisation": functools.partial(
-                lizard_connector.queries.organisation, endpoint="levees")
-        },
-        "leveesections": {
-            "organisation": functools.partial(
-                lizard_connector.queries.organisation, endpoint="leveesections"
-            )
-        },
-        "leveezones": {
-            "organisation": functools.partial(
-                lizard_connector.queries.organisation, endpoint="leveezones")
-        }
-    }