+++ lizard_connector/queries.py
@@ -1,4 +1,6 @@
 # coding=utf-8
+from __future__ import unicode_literals
+
 import datetime
 try:
     import urllib.parse as urlparse
@@ -36,7 +38,7 @@ class QueryDictionary(dict):
         for k, v in f.items():
             self[k] = v
         queries = list(queries)
-        if E and isinstance(E, str):
+        if E and (isinstance(E, str) or isinstance(E, unicode)):
             for k, v in urlparse.parse_qsl(E.strip('?')):
                 self[k] = v
         elif E: